Restructuring code for api writing
